Improve code prettifier hookup to be location independent.
It's a bit tedious to link to a top-level doc-files directory from within
generated documentation, as javadoc doesn't provide any easy way to determine
how deep in the subdirectory structure you are.  However, we can use javascript
to fetch the link to the top-level stylesheet, and then work relative to that.